+++ android/app/src/main/java/com/dgxspark/tongyilite/MainActivity.kt
@@ -7,6 +7,7 @@
 
 package com.dgxspark.tongyilite
 
+import android.content.Intent
 import android.os.Bundle
 import android.os.Handler
 import android.os.Looper